refactor(migration): absorb toast lifecycle and check logic into dialog
Move toast creation (showPendingToast) and data probe logic (check) from
check-database-migration.ts into migration-dialog.svelte.ts. The wrapper
file is now a one-liner delegating to migrationDialog.check(). Removed
isPending setter and migrationToastId getter/setter from public API since
the full toast lifecycle is now internal.
